app/requirements.txt
discord.py
docker
sqlalchemy
psycopg2
python-dotenv